Exception Handling

+++ src/main/java/com/vrzada/webservices/services/UserService.java
@@ -4,34 +4,44 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.vrzada.webservices.entities.User;
 import com.vrzada.webservices.repositories.UserRepository;
+import com.vrzada.webservices.services.exceptions.DatabaseException;
 import com.vrzada.webservices.services.exceptions.ResourceNotFoundException;
 
-@Service //registra a classe como um componente de serviço do Spring
+@Service // registra a classe como um componente de serviço do Spring
 public class UserService {
 	@Autowired
 	private UserRepository repository;
-	
-	public List<User> findAll(){
+
+	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
+		try {
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+
 		repository.deleteById(id);
+		}catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
@@ -42,6 +52,6 @@ public class UserService {
 		entity.setName(obj.getName());
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
-		
+
 	}
 }
